+++ ava.config.js
@@ -1,4 +1,5 @@
 export default {
+  timeout: '1m',
   require: [
     "@babel/register"
   ]