+++ test/test.js
@@ -6,6 +6,12 @@ const processing = (input, options) => {
   return postcss([plugin(options)]).process(input).css;
 };
 
+test('it change circular reference witch declarationByWalk option', t => {
+  const expected = ':root { --nested-var: calc(2 * var(--original-var)); } .class-name { @mixin mixinName calc(2 * 4px); } :root { --original-var: 4px; }';
+  const value = ':root { --nested-var: calc(2 * var(--original-var)); } .class-name { @mixin mixinName var(--nested-var); } :root { --original-var: 4px; }';
+  t.is(processing(value, {declarationByWalk: true}), expected);
+});
+
 test('it change circular reference', t => {
   const expected = ':root{ --from: 1; --to: var(--from)} @for $i from 1 to 1';
   const value = ':root{ --from: 1; --to: var(--from)} @for $i from var(--from) to var(--to)';